Acesso e Configuração Local do Postgres + Fix iews.py close #143
