+++ pypro/base/models.py
@@ -5,6 +5,7 @@ from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
+
 # Create your models here.
 class UserManager(BaseUserManager):
     use_in_migrations = True
